# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Agentic Coder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /project/analyze:
    post:
      operationId: analyzeProject
      tags: [project]
      summary: Analyze project structure and produce a summary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectAnalyzeRequest"
      responses:
        "200":
          description: Project summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
  /project/relevant:
    post:
      operationId: findRelevantFiles
      tags: [project]
      summary: Find files most relevant to a task description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelevantFilesRequest"
      responses:
        "200":
          description: Relevant files ranked by relevance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RelevantFile"
  /project/workspaces:
    get:
      operationId: listWorkspaces
      tags: [project]
      summary: List configured project workspaces
      responses:
        "200":
          description: Workspaces list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: addWorkspace
      tags: [project]
      summary: Add a workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "201":
          description: Workspace added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /project/workspaces/{workspaceId}:
    delete:
      operationId: removeWorkspace
      tags: [project]
      summary: Remove a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
components:
  schemas:
...
    ProjectAnalyzeRequest:
      type: object
      properties:
        path:
          type: string
        includeStats:
          type: boolean
      required:
        - path
    ProjectSummary:
      type: object
      properties:
        path:
          type: string
        name:
          type: string
        fileCount:
          type: integer
        totalSize:
          type: integer
        languages:
          type: array
          items:
            $ref: "#/components/schemas/LanguageStat"
        keyFiles:
          type: array
          items:
            type: string
        gitBranch:
          type: ["string", "null"]
        hasGit:
          type: boolean
      required:
        - path
        - name
        - fileCount
        - totalSize
        - languages
        - keyFiles
        - hasGit
    LanguageStat:
      type: object
      properties:
        language:
          type: string
        fileCount:
          type: integer
        percentage:
          type: number
      required:
        - language
        - fileCount
        - percentage
    RelevantFilesRequest:
      type: object
      properties:
        rootPath:
          type: string
        task:
          type: string
        maxFiles:
          type: integer
      required:
        - rootPath
        - task
    RelevantFile:
      type: object
      properties:
        path:
          type: string
        score:
          type: number
        reason:
          type: string
        size:
          type: integer
      required:
        - path
        - score
        - reason
        - size
    Workspace:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        path:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - path
        - createdAt
    WorkspaceInput:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
      required:
        - name
        - path