# lib/api-zod/src/generated/types/workspace.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Agentic Coder API
 * OpenAPI spec version: 0.1.0
 */

export interface Workspace {
  id: number;
  name: string;
  path: string;
  createdAt: string;
}